+++ Expo1900/Expo1900/Extension/UIView +.swift
@@ -1,0 +1,14 @@
+//
+//  UIView +.swift
+//  Expo1900
+//
+//  Created by kakao on 2022/01/11.
+//
+
+import UIKit
+
+extension UIView {
+    static var identifier: String {
+        String(describing: self)
+    }
+}
